+++ src/main/java/it/prova/gestioneordini/service/CategoriaServiceImpl.java
@@ -8,6 +8,7 @@ import it.prova.gestioneordini.dao.EntityManagerUtil;
 import it.prova.gestioneordini.dao.categoria.CategoriaDAO;
 import it.prova.gestioneordini.model.Articolo;
 import it.prova.gestioneordini.model.Categoria;
+import it.prova.gestioneordini.model.Ordine;
 
 public class CategoriaServiceImpl implements CategoriaService {
 
@@ -122,6 +123,11 @@ public class CategoriaServiceImpl implements CategoriaService {
 
 			categoriaDAO.setEntityManager(entityManager);
 
+			Categoria daRimuovere = categoriaDAO.findByIdFetchingArticoli(idCategoria);
+
+			if (daRimuovere.getArticoli().size() > 0)
+				throw new RuntimeException("test FAILLED, ci sono ancora articoli da rimuovere");
+
 			categoriaDAO.delete(categoriaDAO.get(idCategoria));
 
 			entityManager.getTransaction().commit();
@@ -150,10 +156,27 @@ public class CategoriaServiceImpl implements CategoriaService {
 			articoloInstance = entityManager.merge(articoloInstance);
 
 			articoloInstance.getCategorie().add(categoriaIntsance);
-			
+
 			entityManager.getTransaction().commit();
 		} catch (Exception e) {
 			entityManager.getTransaction().rollback();
+			e.printStackTrace();
+			throw e;
+		} finally {
+			EntityManagerUtil.closeEntityManager(entityManager);
+		}
+	}
+
+	@Override
+	public List<Categoria> cercaCategoriePerOrdini(Categoria categoriaInstance, Ordine ordineInstance)
+			throws Exception {
+		EntityManager entityManager = EntityManagerUtil.getEntityManager();
+		try {
+			categoriaDAO.setEntityManager(entityManager);
+
+			return categoriaDAO.findAllByOrdine(ordineInstance);
+
+		} catch (Exception e) {
 			e.printStackTrace();
 			throw e;
 		} finally {
